refactor(lib): Create lib/features/ for feature-specific modules
- Create lib/features/ directory
- Rename lib/app-nav → lib/features/navigation (2 imports updated)
- Move lib/database/navigation.ts → lib/features/navigation/navigation.ts (3 imports updated)
- Move lib/settings → lib/features/settings (5 imports updated)
- Move lib/office → lib/features/office (1 import updated)
- Fix navigation.ts prisma import to use @/lib/database/client
- Remove empty lib/app-nav directory
- Update all import paths across codebase (11 total imports)
- Rationale: Consolidate navigation modules and group feature-specific functionality
