+++ lib/features/office/blank-document-generator.ts
